+++ apps/question-bank/src/test/java/gov/cdc/nbs/questionbank/CucumberTests.java
@@ -23,7 +23,6 @@ import io.cucumber.spring.CucumberContextConfiguration;
 @CucumberContextConfiguration
 @SpringBootTest
 @ActiveProfiles("test")
-class CucumberTest {
-
+class CucumberTests {
 
 }
